+++ src/backend/utils/adt/arrayfuncs.c
@@ -335,8 +335,8 @@ array_in(PG_FUNCTION_ARGS)
 			if (!AllocSizeIsValid(nbytes))
 				ereturn(escontext, (Datum) 0,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxAllocSize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxAllocSize)));
 		}
 	}
 	if (hasnulls)
@@ -508,8 +508,8 @@ ReadArrayDimensions(char **srcptr, int *ndim_p, int *dim, int *lBound,
 			pg_add_s32_overflow(ub, 1, &ub))
 			ereturn(escontext, false,
 					(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-					 errmsg("array size exceeds the maximum allowed (%d)",
-							(int) MaxArraySize)));
+					 errmsg("array size exceeds the maximum allowed (%zu)",
+							MaxArraySize)));
 
 		dim[ndim] = ub;
 		ndim++;
@@ -741,8 +741,8 @@ ReadArrayStr(char **srcptr,
 					if (maxitems >= MaxArraySize)
 						ereturn(escontext, false,
 								(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-								 errmsg("array size exceeds the maximum allowed (%d)",
-										(int) MaxArraySize)));
+								 errmsg("array size exceeds the maximum allowed (%zu)",
+										MaxArraySize)));
 					maxitems = Min(maxitems * 2, MaxArraySize);
 					values = repalloc_array(values, Datum, maxitems);
 					nulls = repalloc_array(nulls, bool, maxitems);
@@ -1547,8 +1547,8 @@ ReadArrayBinary(StringInfo buf,
 			if (!AllocSizeIsValid(totbytes))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxAllocSize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxAllocSize)));
 		}
 	}
 	*hasnulls = hasnull;
@@ -2359,8 +2359,8 @@ array_set_element(Datum arraydatum,
 				pg_add_s32_overflow(dim[0], addedbefore, &dim[0]))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxArraySize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxArraySize)));
 			lb[0] = indx[0];
 			if (addedbefore > 1)
 				newhasnulls = true; /* will insert nulls */
@@ -2374,8 +2374,8 @@ array_set_element(Datum arraydatum,
 				pg_add_s32_overflow(dim[0], addedafter, &dim[0]))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxArraySize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxArraySize)));
 			if (addedafter > 1)
 				newhasnulls = true; /* will insert nulls */
 		}
@@ -2636,8 +2636,8 @@ array_set_element_expanded(Datum arraydatum,
 				pg_add_s32_overflow(dim[0], addedbefore, &dim[0]))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxArraySize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxArraySize)));
 			lb[0] = indx[0];
 			dimschanged = true;
 			if (addedbefore > 1)
@@ -2652,8 +2652,8 @@ array_set_element_expanded(Datum arraydatum,
 				pg_add_s32_overflow(dim[0], addedafter, &dim[0]))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxArraySize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxArraySize)));
 			dimschanged = true;
 			if (addedafter > 1)
 				newhasnulls = true; /* will insert nulls */
@@ -2913,8 +2913,8 @@ array_set_slice(Datum arraydatum,
 				pg_add_s32_overflow(dim[i], 1, &dim[i]))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxArraySize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxArraySize)));
 
 			lb[i] = lowerIndx[i];
 		}
@@ -2967,8 +2967,8 @@ array_set_slice(Datum arraydatum,
 				pg_add_s32_overflow(dim[0], addedbefore, &dim[0]))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxArraySize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxArraySize)));
 			lb[0] = lowerIndx[0];
 			if (addedbefore > 1)
 				newhasnulls = true; /* will insert nulls */
@@ -2982,8 +2982,8 @@ array_set_slice(Datum arraydatum,
 				pg_add_s32_overflow(dim[0], addedafter, &dim[0]))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxArraySize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxArraySize)));
 			if (addedafter > 1)
 				newhasnulls = true; /* will insert nulls */
 		}
@@ -3323,8 +3323,8 @@ array_map(Datum arrayd,
 			if (!AllocSizeIsValid(nbytes))
 				ereport(ERROR,
 						(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-						 errmsg("array size exceeds the maximum allowed (%d)",
-								(int) MaxAllocSize)));
+						 errmsg("array size exceeds the maximum allowed (%zu)",
+								MaxAllocSize)));
 		}
 	}
 
@@ -3563,8 +3563,8 @@ construct_md_array(Datum *elems,
 		if (!AllocSizeIsValid(nbytes))
 			ereport(ERROR,
 					(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-					 errmsg("array size exceeds the maximum allowed (%d)",
-							(int) MaxAllocSize)));
+					 errmsg("array size exceeds the maximum allowed (%zu)",
+							MaxAllocSize)));
 	}
 
 	/* Allocate and initialize result array */
@@ -5395,8 +5395,8 @@ accumArrayResult(ArrayBuildState *astate,
 		if (!AllocSizeIsValid(astate->alen * sizeof(Datum)))
 			ereport(ERROR,
 					(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-					 errmsg("array size exceeds the maximum allowed (%d)",
-							(int) MaxAllocSize)));
+					 errmsg("array size exceeds the maximum allowed (%zu)",
+							MaxAllocSize)));
 		astate->dvalues = (Datum *)
 			repalloc(astate->dvalues, astate->alen * sizeof(Datum));
 		astate->dnulls = (bool *)
@@ -6234,8 +6234,8 @@ array_fill_internal(ArrayType *dims, ArrayType *lbs,
 			!AllocSizeIsValid(totbytes))
 			ereport(ERROR,
 					(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-					 errmsg("array size exceeds the maximum allowed (%d)",
-							(int) MaxAllocSize)));
+					 errmsg("array size exceeds the maximum allowed (%zu)",
+							MaxAllocSize)));
 
 		/*
 		 * This addition can't overflow, but it might cause us to go past
@@ -6578,8 +6578,8 @@ array_replace_internal(ArrayType *array,
 				if (!AllocSizeIsValid(nbytes))
 					ereport(ERROR,
 							(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-							 errmsg("array size exceeds the maximum allowed (%d)",
-									(int) MaxAllocSize)));
+							 errmsg("array size exceeds the maximum allowed (%zu)",
+									MaxAllocSize)));
 			}
 			nresult++;
 		}
